# backend/internal/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for managing durian farm plots, buyers, sales, maintenance logs, and dashboard analytics.",
        "title": "Durian Farm Management API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/dashboard": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns dashboard overview metrics, recent sales, top plots, and upcoming maintenance for the current user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Get dashboard overview",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User UUID",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD)",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.DashboardResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/dto.ErrorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/dto.ErrorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/dto.ErrorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/dashboard/activity": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns recent sales and upcoming maintenance activity for the current user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Get dashboard activity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User UUID",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD)",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.DashboardActivityResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/dto.ErrorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/dto.ErrorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/dto.ErrorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/dashboard/revenue": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns revenue summary analytics for the current user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Get dashboard revenue",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User UUID",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD)",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.SalesSummaryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/dto.ErrorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/dto.ErrorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/dto.ErrorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/dto.ErrorResponse"
                },
                "meta": {
                    "$ref": "#/definitions/dto.PaginationMeta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.DashboardActivityResponse": {
            "type": "object",
            "properties": {
                "recent_sales": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.DashboardRecentSaleResponse"
                    }
                },
                "upcoming_maintenance": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.DashboardUpcomingMaintenanceRecord"
                    }
                }
            }
        },
        "dto.DashboardComparisonResponse": {
            "type": "object",
            "properties": {
                "revenue_change_percent": {
                    "type": "number"
                },
                "weight_change_percent": {
                    "type": "number"
                }
            }
        },
        "dto.DashboardGradeResponse": {
            "type": "object",
            "properties": {
                "percentage": {
                    "type": "number"
                },
                "weight_kg": {
                    "type": "number"
                }
            }
        },
        "dto.DashboardRecentSaleResponse": {
            "type": "object",
            "properties": {
                "grade": {
                    "$ref": "#/definitions/models.Grade"
                },
                "id": {
                    "type": "string"
                },
                "plot_name": {
                    "type": "string"
                },
                "sale_date": {
                    "type": "string"
                },
                "total_price": {
                    "type": "number"
                },
                "weight_kg": {
                    "type": "number"
                }
            }
        },
        "dto.DashboardResponse": {
            "type": "object",
            "properties": {
                "comparison": {
                    "$ref": "#/definitions/dto.DashboardComparisonResponse"
                },
                "grade_distribution": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/dto.DashboardGradeResponse"
                    }
                },
                "period": {
                    "$ref": "#/definitions/dto.PeriodResponse"
                },
                "recent_sales": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.DashboardRecentSaleResponse"
                    }
                },
                "stats": {
                    "$ref": "#/definitions/dto.DashboardStatsResponse"
                },
                "top_plots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.DashboardTopPlotResponse"
                    }
                },
                "upcoming_maintenance": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.DashboardUpcomingMaintenanceRecord"
                    }
                }
            }
        },
        "dto.DashboardStatsResponse": {
            "type": "object",
            "properties": {
                "average_price_per_kg": {
                    "type": "number"
                },
                "total_revenue": {
                    "type": "number"
                },
                "total_weight_kg": {
                    "type": "number"
                },
                "transaction_count": {
                    "type": "integer"
                }
            }
        },
        "dto.DashboardTopPlotResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "revenue": {
                    "type": "number"
                }
            }
        },
        "dto.DashboardUpcomingMaintenanceRecord": {
            "type": "object",
            "properties": {
                "activity_date": {
                    "type": "string"
                },
                "activity_type": {
                    "$ref": "#/definitions/models.ActivityType"
                },
                "id": {
                    "type": "string"
                },
                "plot_name": {
                    "type": "string"
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ValidationError"
                    }
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "dto.GradeSummaryResponse": {
            "type": "object",
            "properties": {
                "percentage": {
                    "type": "number"
                },
                "revenue": {
                    "type": "number"
                },
                "weight_kg": {
                    "type": "number"
                }
            }
        },
...
        "dto.PaginationMeta": {
            "type": "object",
            "properties": {
                "has_next": {
                    "type": "boolean"
                },
                "has_prev": {
                    "type": "boolean"
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "dto.PeriodResponse": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                }
            }
        },
...
        "dto.SalesSummaryResponse": {
            "type": "object",
            "properties": {
                "by_grade": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/dto.GradeSummaryResponse"
                    }
                },
                "period": {
                    "$ref": "#/definitions/dto.PeriodResponse"
                },
                "timeline": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SalesTimelinePointResponse"
                    }
                },
                "totals": {
                    "$ref": "#/definitions/dto.SalesTotalsResponse"
                }
            }
        },
        "dto.SalesTimelinePointResponse": {
            "type": "object",
            "properties": {
                "period": {
                    "type": "string"
                },
                "revenue": {
                    "type": "number"
                },
                "transaction_count": {
                    "type": "integer"
                },
                "weight_kg": {
                    "type": "number"
                }
            }
        },
        "dto.SalesTotalsResponse": {
            "type": "object",
            "properties": {
                "average_price_per_kg": {
                    "type": "number"
                },
                "revenue": {
                    "type": "number"
                },
                "transaction_count": {
                    "type": "integer"
                },
                "weight_kg": {
                    "type": "number"
                }
            }
        },
...
        "dto.ValidationError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.ActivityType": {
            "type": "string",
            "enum": [
                "watering",
                "fertilizing",
                "pruning",
                "pest_control",
                "harvesting"
            ],
            "x-enum-varnames": [
                "ActivityWatering",
                "ActivityFertilizing",
                "ActivityPruning",
                "ActivityPestControl",
                "ActivityHarvesting"
            ]
        },
        "models.Grade": {
            "type": "string",
            "enum": [
                "A",
                "B"
            ],
            "x-enum-comments": {
                "GradeA": "Premium quality",
                "GradeB": "Standard quality"
            },
            "x-enum-descriptions": [
                "Premium quality",
                "Standard quality"
            ],
            "x-enum-varnames": [
                "GradeA",
                "GradeB"
            ]
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-User-ID",
            "in": "header"
        }
    }
}